refactor(calling): add multi-browser support and restructure constants
- Add Firefox, Edge, and Safari browser configs via PW_BROWSER env var
- Split constants into separate modules: selectors.ts, timeouts.ts, index.ts
- Move OAuth constants (ENV_PATH, WIDGETS_URL) from oauth.setup.ts to constants
- Remove unused LOCAL_VIDEO and REMOTE_VIDEO selectors
