+++ playwright/calling/constants/selectors.ts
@@ -1,14 +1,3 @@
-export type ServiceIndicator = 'calling' | 'contactcenter' | 'guestcalling';
-
-export const SAMPLE_APP_PATH = '/samples/calling/';
-export const CC_SERVICE_DOMAIN = 'rtw.prod-us1.rtmsprod.net';
-
-// Timeouts — SDK timeout + 5s buffer for network/UI overhead
-export const AWAIT_TIMEOUT = 10000; // General UI interactions
-export const SDK_INIT_TIMEOUT = 65000; // RETRY_TIMER_UPPER_LIMIT (60s) + 5s
-export const REGISTRATION_TIMEOUT = 35000; // BASE_REG_RETRY_TIMER_VAL_IN_SEC (30s) + 5s
-export const OPERATION_TIMEOUT = 15000; // SUPPLEMENTARY_SERVICES_TIMEOUT (10s) + 5s
-
 // Element selectors (from the calling sample app)
 export const SELECTORS = {
   // Authentication
@@ -44,9 +33,7 @@ export const SELECTORS = {
 
   // Media
   GET_MEDIA_STREAMS_BTN: '#sd-get-media-streams',
-  LOCAL_VIDEO: '#local-video',
   LOCAL_AUDIO: '#local-audio',
-  REMOTE_VIDEO: '#remote-video',
   REMOTE_AUDIO: '#remote-audio',
 
   // Guest Calling
